requirements.txt
Flask==3.0.2
flask-socketio
eventlet
pytest>=8

# ─── new for authentication ───────────────────────────
Flask-Login>=0.6
passlib[bcrypt]~=1.7
Flask-WTF~=1.2
Flask-SQLAlchemy~=3.1
Flask-Migrate~=4.0
argon2-cffi
Flask-Mail>=0.9      # or Flask-Mail
pyotp~=2.9
Authlib[client]>=1.3
python-dotenv
twilio~=9.0